@@ -1,19 +1,19 @@
-
 import json
-import os.path as osp
+from pathlib import Path
 
 from ._version import __version__
 
-HERE = osp.abspath(osp.dirname(__file__))
 
-with open(osp.join(HERE, 'labextension', 'package.json')) as fid:
+HERE = Path(__file__).parent.resolve()
+
+
+with (HERE / "labextension" / "package.json").open() as fid:
     data = json.load(fid)
+
 
 def _jupyter_labextension_paths():
     return [{
-        'src': 'labextension',
-        'dest': data['name']
+        "src": "labextension",
+        "dest": data["name"]
     }]
-
-
 
