@@ -6,12 +6,13 @@ import {
 import { IThemeManager } from '@jupyterlab/apputils';
 
 /**
- * A plugin for jupyterlab-theme-solarized-dark
+ * Initialization data for the jupyterlab-theme-solarized-dark extension.
  */
 const plugin: JupyterFrontEndPlugin<void> = {
   id: 'jupyterlab-theme-solarized-dark:plugin',
+  autoStart: true,
   requires: [IThemeManager],
-  activate: function(app: JupyterFrontEnd, manager: IThemeManager) {
+  activate: (app: JupyterFrontEnd, manager: IThemeManager) => {
     const style = 'jupyterlab-theme-solarized-dark/index.css';
 
     manager.register({
@@ -21,8 +22,7 @@ const plugin: JupyterFrontEndPlugin<void> = {
       load: () => manager.loadCSS(style),
       unload: () => Promise.resolve(undefined)
     });
-  },
-  autoStart: true
+  }
 };
 
 export default plugin;
